requirements.txt
attrs==22.1.0
black==22.6.0
click==8.1.3
cycler==0.11.0
fonttools==4.34.4
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
mypy-extensions==0.4.3
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.3.0
